Fixed delay bug and introduced base class for RXTX implementations
