+++ host/build.number
@@ -1,3 +1,3 @@
 #Build Number for ANT. Do not edit!
-#Mon Jan 19 13:19:37 CST 2015
-build.number=38
+#Mon Jan 19 14:17:21 CST 2015
+build.number=39
