+++ host/networkinstall.sh
@@ -3,9 +3,9 @@
 rm -r cwh
 mkdir cwh
 cd cwh
-wget https://github.com/WesGilster/Creation-Workshop-Host/raw/master/host/cwh-0.37.zip
-unzip cwh-0.37.zip
+wget https://github.com/WesGilster/Creation-Workshop-Host/raw/master/host/cwh-0.38.zip
+unzip cwh-0.38.zip
 chmod 777 start.sh
 chmod 777 debug.sh
 chmod 777 browseprinter.sh
-rm cwh-0.37.zip
+rm cwh-0.38.zip
